moved clients under provider. renamed models to model. moved config/util  to top level. removed test
