+++ config/new_config_resources_only.yml
