+++ fabfed/provider/fabric/fabric_slice.py
@@ -30,11 +30,11 @@ from fabrictestbed_extensions.fablib.fablib import fablib
 from fabrictestbed_extensions.fablib.slice import Slice
 from tabulate import tabulate
 
-from fabfed.controller.util.config import Config
-from fabfed.models import AbstractResourceListener
-from fabfed.models import AbstractSlice
+from fabfed.util.config import Config
+from fabfed.model import AbstractResourceListener
+from fabfed.model import AbstractSlice
 from .fabric_node import FabricNode, NodeBuilder
-from .fabric_network import FabricNetwork, NetworkBuilder
+from .fabric_network import NetworkBuilder
 
 
 class FabricSlice(AbstractSlice, AbstractResourceListener):
