+++ tests/test_config.py
@@ -1,4 +1,4 @@
-from fabfed.controller.util.parser import Parser, ParseConfigException, DependencyInfo
+from fabfed.util.parser import Parser, ParseConfigException, DependencyInfo
 import pytest
